+++ hazelcast-client/src/test/java/com/hazelcast/client/reliableidgen/impl/ReliableIdGenerator_NodeIdOverflowIntegrationTest.java
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-package com.hazelcast.concurrent.reliableidgen;
+package com.hazelcast.client.reliableidgen.impl;
 
-import com.hazelcast.core.ReliableIdGenerator;
+import com.hazelcast.client.config.ClientConfig;
+import com.hazelcast.client.test.TestHazelcastFactory;
+import com.hazelcast.reliableidgen.ReliableIdGenerator;
 import com.hazelcast.core.HazelcastException;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.instance.MemberImpl;
 import com.hazelcast.test.HazelcastParallelClassRunner;
-import com.hazelcast.test.TestHazelcastInstanceFactory;
 import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
 import org.junit.After;
@@ -38,7 +39,7 @@ public class ReliableIdGenerator_NodeIdOverflowIntegrationTest {
     @Rule
     public ExpectedException exception = ExpectedException.none();
 
-    private TestHazelcastInstanceFactory factory = new TestHazelcastInstanceFactory();
+    private TestHazelcastFactory factory = new TestHazelcastFactory();
 
     @After
     public void after() {
@@ -47,30 +48,47 @@ public class ReliableIdGenerator_NodeIdOverflowIntegrationTest {
 
     @Test
     public void when_memberOutOfRangeNodeId_then_theOtherMemberUsed() throws Exception {
+        // Design of this test: we create two members. To one of them, out-of-range memberListJoinVersion
+        // is assigned. This causes the client message to fail and the client will retry with another member
+        // and will never again try the failed member.
+        //
+        // Probability of initially choosing a good member and not going through the process of choosing another
+        // one is 50%. To avoid false success, we retry 10 times, which gives the probability of false success
+        // of 0.1%.
         HazelcastInstance instance1 = factory.newHazelcastInstance();
         HazelcastInstance instance2 = factory.newHazelcastInstance();
-        assignOutOfRangeNodeId(instance2);
+        assignOverflowedNodeId(instance2);
 
-        // let's use the instance with out-of-range node ID to generate IDs, it should succeed
-        ReliableIdGenerator gen = instance2.getReliableIdGenerator("gen");
-        gen.newId();
+        ClientConfig clientConfig = new ClientConfig();
+        // disable smart routing - such clients must also work reliably
+        clientConfig.getNetworkConfig().setSmartRouting(false);
+        for (int i = 0; i < 10; i++) {
+            System.out.println("Creating client " + i);
+            HazelcastInstance client = factory.newHazelcastClient(clientConfig);
+            ReliableIdGenerator gen = client.getReliableIdGenerator("gen");
+            for (int j = 0; j < 100; j++) {
+                // call should not fail
+                gen.newId();
+            }
+        }
     }
 
     @Test
     public void when_allMembersOutOfRangeNodeId_then_error() {
         HazelcastInstance instance1 = factory.newHazelcastInstance();
         HazelcastInstance instance2 = factory.newHazelcastInstance();
-        assignOutOfRangeNodeId(instance1);
-        assignOutOfRangeNodeId(instance2);
+        assignOverflowedNodeId(instance1);
+        assignOverflowedNodeId(instance2);
 
-        ReliableIdGenerator gen = instance1.getReliableIdGenerator("gen");
+        HazelcastInstance client = factory.newHazelcastClient();
+        ReliableIdGenerator gen = client.getReliableIdGenerator("gen");
 
         exception.expect(HazelcastException.class);
         exception.expectMessage("All members have node ID out of range");
         gen.newId();
     }
 
-    private void assignOutOfRangeNodeId(HazelcastInstance instance2) {
+    private void assignOverflowedNodeId(HazelcastInstance instance2) {
         MemberImpl member = (MemberImpl) instance2.getCluster().getLocalMember();
         member.setMemberListJoinVersion(100000);
     }
