+++ hazelcast/src/main/java/com/hazelcast/cache/CacheEventType.java
@@ -66,7 +66,7 @@ public enum CacheEventType {
     EXPIRATION_TIME_UPDATED(8),
 
     /**
-     * An event type indicating that partition lost is detected in given cache with name
+     * An event type indicating that partition loss is detected in given cache with name
      */
     PARTITION_LOST(9);
