+++ hazelcast/src/main/java/com/hazelcast/internal/cluster/impl/ClusterServiceImpl.java
@@ -888,7 +888,7 @@ public class ClusterServiceImpl implements ClusterService, ConnectionListener, M
                 throw new IllegalStateException("Member list join version is not available when not joined");
             } else if (getClusterVersion().isLessThan(V3_10)) {
                 String msg = "Member list join version is not available with a cluster version less than 3.10";
-                throw new IllegalStateException(msg);
+                throw new UnsupportedOperationException(msg);
             }
 
             int joinVersion = localMember.getMemberListJoinVersion();
