+++ hazelcast/src/main/java/com/hazelcast/reliableidgen/ReliableIdGenerator.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package com.hazelcast.core;
+package com.hazelcast.reliableidgen;
 
-import com.hazelcast.concurrent.reliableidgen.NodeIdOutOfRangeException;
+import com.hazelcast.reliableidgen.impl.NodeIdOutOfRangeException;
+import com.hazelcast.core.DistributedObject;
 import com.hazelcast.internal.cluster.ClusterService;
 
 /**
@@ -29,7 +30,7 @@ import com.hazelcast.internal.cluster.ClusterService;
  * members, which makes the generator safe even in split-brain scenario (for caveats,
  * {@link ClusterService#getMemberListJoinVersion() see here}).
  * <p>
- * Timestamp component is in milliseconds since 1.1.2017, 0:00 GMT, and has 42 bits. This caps
+ * Timestamp component is in milliseconds since 1.1.2017, 0:00 UTC, and has 42 bits. This caps
  * the useful lifespan of the generator to little less than 140 years. The sequence component is 6 bits.
  * If more than 64 IDs are requested in single millisecond, IDs will gracefully overflow to next
  * millisecond and uniqueness is guaranteed in this case.
@@ -48,6 +49,8 @@ import com.hazelcast.internal.cluster.ClusterService;
  * member with join version smaller than 2^16 in the cluster. The remedy is to restart the cluster:
  * nodeId will be assigned from zero again. Uniqueness after the restart will be preserved thanks to
  * the timestamp component.
+ *
+ * @since 3.10
  */
 public interface ReliableIdGenerator extends DistributedObject {
 
@@ -62,7 +65,8 @@ public interface ReliableIdGenerator extends DistributedObject {
      * @return new cluster-wide unique ID
      *
      * @throws NodeIdOutOfRangeException if node ID for all members in the cluster is out of valid range.
-     * See "Node ID overflow" in {@link ReliableIdGenerator class documentation} for more details.
+     *      See "Node ID overflow" in {@link ReliableIdGenerator class documentation} for more details.
+     * @throws UnsupportedOperationException if the cluster version is below 3.10
      */
     long newId();
 }
