+++ hazelcast/src/main/java/com/hazelcast/reliableidgen/impl/IdBatch.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.hazelcast.concurrent.reliableidgen;
+package com.hazelcast.reliableidgen.impl;
 
-import com.hazelcast.core.ReliableIdGenerator;
+import com.hazelcast.reliableidgen.ReliableIdGenerator;
 
 import javax.annotation.Nonnull;
 import java.util.Iterator;
