+++ hazelcast/src/main/resources/META-INF/services/com.hazelcast.DataSerializerHook
@@ -31,4 +31,4 @@ com.hazelcast.aggregation.impl.AggregatorDataSerializerHook
 com.hazelcast.projection.impl.ProjectionDataSerializerHook
 com.hazelcast.config.ConfigDataSerializerHook
 com.hazelcast.journal.EventJournalDataSerializerHook
-com.hazelcast.concurrent.reliableidgen.ReliableIdGeneratorDataSerializerHook
+com.hazelcast.reliableidgen.impl.ReliableIdGeneratorDataSerializerHook
