+++ hazelcast/src/test/java/com/hazelcast/reliableidgen/impl/ReliableIdGenerator_IdBatchTest.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package com.hazelcast.concurrent.reliableidgen;
+package com.hazelcast.reliableidgen.impl;
 
+import com.hazelcast.reliableidgen.impl.IdBatch;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
