+++ hazelcast/src/test/java/com/hazelcast/reliableidgen/impl/ReliableIdGenerator_NodeIdOverflowIntegrationTest.java
@@ -14,15 +14,14 @@
  * limitations under the License.
  */
 
-package com.hazelcast.client.reliableidgen;
+package com.hazelcast.reliableidgen.impl;
 
-import com.hazelcast.client.config.ClientConfig;
-import com.hazelcast.client.test.TestHazelcastFactory;
-import com.hazelcast.core.ReliableIdGenerator;
+import com.hazelcast.reliableidgen.ReliableIdGenerator;
 import com.hazelcast.core.HazelcastException;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.instance.MemberImpl;
 import com.hazelcast.test.HazelcastParallelClassRunner;
+import com.hazelcast.test.TestHazelcastInstanceFactory;
 import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
 import org.junit.After;
@@ -39,7 +38,7 @@ public class ReliableIdGenerator_NodeIdOverflowIntegrationTest {
     @Rule
     public ExpectedException exception = ExpectedException.none();
 
-    private TestHazelcastFactory factory = new TestHazelcastFactory();
+    private TestHazelcastInstanceFactory factory = new TestHazelcastInstanceFactory();
 
     @After
     public void after() {
@@ -48,47 +47,30 @@ public class ReliableIdGenerator_NodeIdOverflowIntegrationTest {
 
     @Test
     public void when_memberOutOfRangeNodeId_then_theOtherMemberUsed() throws Exception {
-        // Design of this test: we create two members. To one of them, out-of-range memberListJoinVersion
-        // is assigned. This causes the client message to fail and the client will retry with another member
-        // and will never again try the failed member.
-        //
-        // Probability of initially choosing a good member and not going through the process of choosing another
-        // one is 50%. To avoid false success, we retry 10 times, which gives the probability of false success
-        // of 0.1%.
         HazelcastInstance instance1 = factory.newHazelcastInstance();
         HazelcastInstance instance2 = factory.newHazelcastInstance();
-        assignOverflowedNodeId(instance2);
+        assignOutOfRangeNodeId(instance2);
 
-        ClientConfig clientConfig = new ClientConfig();
-        // disable smart routing - such clients must also work reliably
-        clientConfig.getNetworkConfig().setSmartRouting(false);
-        for (int i = 0; i < 10; i++) {
-            System.out.println("Creating client " + i);
-            HazelcastInstance client = factory.newHazelcastClient(clientConfig);
-            ReliableIdGenerator gen = client.getReliableIdGenerator("gen");
-            for (int j = 0; j < 100; j++) {
-                // call should not fail
-                gen.newId();
-            }
-        }
+        // let's use the instance with out-of-range node ID to generate IDs, it should succeed
+        ReliableIdGenerator gen = instance2.getReliableIdGenerator("gen");
+        gen.newId();
     }
 
     @Test
     public void when_allMembersOutOfRangeNodeId_then_error() {
         HazelcastInstance instance1 = factory.newHazelcastInstance();
         HazelcastInstance instance2 = factory.newHazelcastInstance();
-        assignOverflowedNodeId(instance1);
-        assignOverflowedNodeId(instance2);
+        assignOutOfRangeNodeId(instance1);
+        assignOutOfRangeNodeId(instance2);
 
-        HazelcastInstance client = factory.newHazelcastClient();
-        ReliableIdGenerator gen = client.getReliableIdGenerator("gen");
+        ReliableIdGenerator gen = instance1.getReliableIdGenerator("gen");
 
         exception.expect(HazelcastException.class);
         exception.expectMessage("All members have node ID out of range");
         gen.newId();
     }
 
-    private void assignOverflowedNodeId(HazelcastInstance instance2) {
+    private void assignOutOfRangeNodeId(HazelcastInstance instance2) {
         MemberImpl member = (MemberImpl) instance2.getCluster().getLocalMember();
         member.setMemberListJoinVersion(100000);
     }
